+++ testing/steps/test_step_cloner_for_each_data_input.py
@@ -1,9 +1,8 @@
-from typing import Any
-
 from neuraxle.hyperparams.distributions import Boolean
 from neuraxle.hyperparams.space import HyperparameterSpace, HyperparameterSamples
 from neuraxle.steps.util import StepClonerForEachDataInput
 from testing.test_pipeline import SomeStep
+
 
 HYPE_SPACE = HyperparameterSpace({
     "a__test": Boolean()
@@ -14,22 +13,6 @@ HYPE_SAMPLE = HyperparameterSamples({
 })
 
 
-def test_step_cloner_should_set_wrapped_step_hyperparams():
-    step_cloner = StepClonerForEachDataInput(SomeStep())
-
-    step_cloner.set_hyperparams(HYPE_SAMPLE)
-
-    assert step_cloner.step.get_hyperparams() == HYPE_SAMPLE
-
-
-def test_step_cloner_should_set_wrapped_step_hyperparams_space():
-    step_cloner = StepClonerForEachDataInput(SomeStep())
-
-    step_cloner.set_hyperparams_space(HYPE_SPACE)
-
-    assert step_cloner.step.get_hyperparams_space() == HYPE_SPACE
-
-
 class SomeStepInverseTransform(SomeStep):
     def fit_transform(self, data_inputs, expected_outputs=None):
         return self, 'fit_transform'
@@ -38,7 +21,7 @@ class SomeStepInverseTransform(SomeStep):
         return 'inverse_transform'
 
 
-def test_step_cloner_should_fit_transform():
+def test_should_fit_transform():
     some_step = SomeStepInverseTransform()
     step_cloner = StepClonerForEachDataInput(some_step)
 
@@ -48,7 +31,7 @@ def test_step_cloner_should_fit_transform():
     assert processed_outputs == ['fit_transform']
 
 
-def test_step_cloner_should_inverse_transform():
+def test_should_inverse_transform():
     step_cloner = StepClonerForEachDataInput(SomeStepInverseTransform())
 
     step_cloner, processed_outputs = step_cloner.fit_transform([0])
